+++ packages/middlewares/src/types.ts
@@ -1,0 +1,12 @@
+import { NextMiddlewareResult } from 'next/dist/server/web/types'
+import { NextFetchEvent, NextRequest, NextResponse } from 'next/server'
+
+export type CustomMiddleware = (
+  request: NextRequest,
+  event: NextFetchEvent,
+  response: NextResponse,
+) => NextMiddlewareResult | Promise<NextMiddlewareResult>
+
+export type MiddlewareFactory = (
+  middleware: CustomMiddleware,
+) => CustomMiddleware
